# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Food Tour API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VenueNameLocal = { [key: string]: string };

export type VenueCategory = (typeof VenueCategory)[keyof typeof VenueCategory];

export const VenueCategory = {
  vietnamese: "vietnamese",
  seafood: "seafood",
  vegetarian: "vegetarian",
  hotpot: "hotpot",
  "banh-mi": "banh-mi",
  coffee: "coffee",
  other: "other",
} as const;

export type VenuePriceRange =
  (typeof VenuePriceRange)[keyof typeof VenuePriceRange];

export const VenuePriceRange = {
  $: "$",
  $$: "$$",
  $$$: "$$$",
} as const;

export interface Venue {
  id: string;
  name: string;
  nameLocal?: VenueNameLocal;
  category: VenueCategory;
  description?: string;
  lat: number;
  lng: number;
  address: string;
  rating: number;
  reviewCount?: number;
  imageUrl?: string;
  isOpen: boolean;
  /** Trigger radius in meters */
  audioRadius: number;
  priceRange?: VenuePriceRange;
  tags?: string[];
}

export type VenueDetailDescriptionLocal = { [key: string]: string };

export type VenueDetailHours = { [key: string]: string };

export type MenuItemNameLocal = { [key: string]: string };

export interface MenuItem {
  id: string;
  name: string;
  nameLocal?: MenuItemNameLocal;
  description?: string;
  price: number;
  imageUrl?: string;
  isFeatured?: boolean;
}

export interface Review {
  id: string;
  author: string;
  rating: number;
  comment: string;
  date: string;
  lang?: string;
}

export type VenueDetail = Venue & {
  descriptionLocal?: VenueDetailDescriptionLocal;
  phone?: string;
  website?: string;
  hours?: VenueDetailHours;
  menu?: MenuItem[];
  gallery?: string[];
  reviews?: Review[];
  hasAudio?: boolean;
  audioLanguages?: string[];
};

export type NearbyVenue = Venue & {
  /** Distance in meters */
  distance?: number;
  withinAudioRadius?: boolean;
};

export interface Language {
  code: string;
  name: string;
  nativeName: string;
  flag: string;
}

export interface ChatRequest {
  message: string;
  lang: string;
  userLat?: number;
  userLng?: number;
  /** Optional venue context */
  context?: string;
}

export interface ChatResponse {
  reply: string;
  suggestedVenues?: Venue[];
  quickReplies?: string[];
}

export interface AudioInfo {
  venueId: string;
  lang: string;
  audioUrl: string;
  fallbackLang?: string;
  transcript?: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetVenuesParams = {
  /**
   * Language code for content
   */
  lang?: string;
  /**
   * Filter by food category
   */
  category?: string;
  /**
   * User latitude for distance sorting
   */
  lat?: number;
  /**
   * User longitude for distance sorting
   */
  lng?: number;
};

export type GetVenueByIdParams = {
  lang?: string;
};

export type GetNearbyVenuesParams = {
  lat: number;
  lng: number;
  /**
   * Radius in meters (default 100)
   */
  radius?: number;
  lang?: string;
};

export type GetAudioForVenueParams = {
  /**
   * Language code (default en)
   */
  lang?: string;
};