# lib/api-zod/src/generated/types/audioInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Food Tour API
 * OpenAPI spec version: 0.1.0
 */

export interface AudioInfo {
  venueId: string;
  lang: string;
  audioUrl: string;
  fallbackLang?: string;
  transcript?: string;
}