# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Food Tour API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatRequest {
  message: string;
  lang: string;
  userLat?: number;
  userLng?: number;
  /** Optional venue context */
  context?: string;
}