# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Food Tour API
 * OpenAPI spec version: 0.1.0
 */
import type { Venue } from "./venue";

export interface ChatResponse {
  reply: string;
  suggestedVenues?: Venue[];
  quickReplies?: string[];
}