# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Food Tour API
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItemNameLocal } from "./menuItemNameLocal";

export interface MenuItem {
  id: string;
  name: string;
  nameLocal?: MenuItemNameLocal;
  description?: string;
  price: number;
  imageUrl?: string;
  isFeatured?: boolean;
}