# lib/api-zod/src/generated/types/nearbyVenue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Food Tour API
 * OpenAPI spec version: 0.1.0
 */
import type { Venue } from "./venue";

export type NearbyVenue = Venue & {
  /** Distance in meters */
  distance?: number;
  withinAudioRadius?: boolean;
};