# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Food Tour API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: string;
  author: string;
  rating: number;
  comment: string;
  date: string;
  lang?: string;
}