# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Food Tour API
 * OpenAPI spec version: 0.1.0
 */
import type { VenueCategory } from "./venueCategory";
import type { VenueNameLocal } from "./venueNameLocal";
import type { VenuePriceRange } from "./venuePriceRange";

export interface Venue {
  id: string;
  name: string;
  nameLocal?: VenueNameLocal;
  category: VenueCategory;
  description?: string;
  lat: number;
  lng: number;
  address: string;
  rating: number;
  reviewCount?: number;
  imageUrl?: string;
  isOpen: boolean;
  /** Trigger radius in meters */
  audioRadius: number;
  priceRange?: VenuePriceRange;
  tags?: string[];
}